requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pypdf
python-docx
